move Vtk functionality to own module
