+++ src/CLIMA.jl
@@ -15,6 +15,6 @@ include("DGmethods/DGBalanceLawDiscretizations.jl")
 include("ODESolvers/ODESolvers.jl")
 include("ODESolvers/LowStorageRungeKuttaMethod.jl")
 include("ODESolvers/GenericCallbacks.jl")
-
+include("InputOutput/Vtk/Vtk.jl")
 
 end
