+++ test/Atmos/Parameterizations/TurbulenceConvection/runtests.jl
@@ -1,10 +1,5 @@
 using Test
 
-for submodule in ["Grids",
-                  "StateVecs",
-                  "StateVecFuncs",
-                  ]
-
-  println("Testing $submodule")
-  include(joinpath("../test",submodule*".jl"))
-end
+include("Grids.jl")
+include("StateVecs.jl")
+include("StateVecFuncs.jl")
